chore: move executeDeploy logic from CLI to SDK (#6222)
### Description

<!--
What's included in this PR?
-->

- Move logic from `executeDeploy` that was in the `CLI` to the `SDK`,
now this function is named `executeWarpDeploy`
- Instead of using the CLI context, the function will receive the
parameters

### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->
No
### Related issues

<!--
- Fixes #[issue number here]
-->
fix #6138 
### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->
Yes
### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->
Manual local test
