+++ .changeset/orange-women-show.md
@@ -1,0 +1,6 @@
+---
+'@hyperlane-xyz/cli': minor
+'@hyperlane-xyz/sdk': minor
+---
+
+Move executeDeploy logic from CLI to SDK
