+++ typescript/sdk/src/deploy/warp.ts
@@ -1,0 +1,215 @@
+import { ProxyAdmin__factory } from '@hyperlane-xyz/core';
+import { buildArtifact as coreBuildArtifact } from '@hyperlane-xyz/core/buildArtifact.js';
+import { ChainAddresses } from '@hyperlane-xyz/registry';
+import {
+  Address,
+  objMap,
+  promiseObjAll,
+  rootLogger,
+} from '@hyperlane-xyz/utils';
+
+import { CCIPContractCache } from '../ccip/utils.js';
+import { HyperlaneContractsMap } from '../contracts/types.js';
+import { EvmHookModule } from '../hook/EvmHookModule.js';
+import { HookConfig } from '../hook/types.js';
+import { EvmIsmModule } from '../ism/EvmIsmModule.js';
+import { IsmConfig } from '../ism/types.js';
+import { MultiProvider } from '../providers/MultiProvider.js';
+import { HypERC20Factories, HypERC721Factories } from '../token/contracts.js';
+import { HypERC20Deployer, HypERC721Deployer } from '../token/deploy.js';
+import {
+  HypTokenRouterConfig,
+  WarpRouteDeployConfigMailboxRequired,
+} from '../token/types.js';
+import { ChainMap } from '../types.js';
+import { extractIsmAndHookFactoryAddresses } from '../utils/ism.js';
+
+import { HyperlaneProxyFactoryDeployer } from './HyperlaneProxyFactoryDeployer.js';
+import { ContractVerifier } from './verify/ContractVerifier.js';
+import { ExplorerLicenseType } from './verify/types.js';
+
+export async function executeWarpDeploy(
+  multiProvider: MultiProvider,
+  warpDeployConfig: WarpRouteDeployConfigMailboxRequired,
+  registryAddresses: ChainMap<ChainAddresses>,
+  apiKeys: ChainMap<string>,
+): Promise<HyperlaneContractsMap<HypERC20Factories | HypERC721Factories>> {
+  const deployer = warpDeployConfig.isNft
+    ? new HypERC721Deployer(multiProvider)
+    : new HypERC20Deployer(multiProvider); // TODO: replace with EvmERC20WarpModule
+
+  const contractVerifier = new ContractVerifier(
+    multiProvider,
+    apiKeys,
+    coreBuildArtifact,
+    ExplorerLicenseType.MIT,
+  );
+
+  const ismFactoryDeployer = new HyperlaneProxyFactoryDeployer(
+    multiProvider,
+    contractVerifier,
+  );
+
+  // For each chain in WarpRouteConfig, deploy each Ism Factory, if it's not in the registry
+  // Then return a modified config with the ism and/or hook address as a string
+  const modifiedConfig = await resolveWarpIsmAndHook(
+    warpDeployConfig,
+    multiProvider,
+    registryAddresses,
+    ismFactoryDeployer,
+    contractVerifier,
+  );
+
+  const deployedContracts = await deployer.deploy(modifiedConfig);
+
+  return deployedContracts;
+}
+
+async function resolveWarpIsmAndHook(
+  warpConfig: WarpRouteDeployConfigMailboxRequired,
+  multiProvider: MultiProvider,
+  registryAddresses: ChainMap<ChainAddresses>,
+  ismFactoryDeployer: HyperlaneProxyFactoryDeployer,
+  contractVerifier?: ContractVerifier,
+): Promise<WarpRouteDeployConfigMailboxRequired> {
+  return promiseObjAll(
+    objMap(warpConfig, async (chain, config) => {
+      const ccipContractCache = new CCIPContractCache(registryAddresses);
+      const chainAddresses = registryAddresses[chain];
+
+      if (!chainAddresses) {
+        throw `Registry factory addresses not found for ${chain}.`;
+      }
+
+      config.interchainSecurityModule = await createWarpIsm({
+        ccipContractCache,
+        chain,
+        chainAddresses,
+        multiProvider,
+        contractVerifier,
+        ismFactoryDeployer,
+        warpConfig: config,
+      }); // TODO write test
+
+      config.hook = await createWarpHook({
+        ccipContractCache,
+        chain,
+        chainAddresses,
+        multiProvider,
+        contractVerifier,
+        ismFactoryDeployer,
+        warpConfig: config,
+      });
+      return config;
+    }),
+  );
+}
+
+/**
+ * Deploys the Warp ISM for a given config
+ *
+ * @returns The deployed ism address
+ */
+async function createWarpIsm({
+  ccipContractCache,
+  chain,
+  chainAddresses,
+  multiProvider,
+  contractVerifier,
+  warpConfig,
+}: {
+  ccipContractCache: CCIPContractCache;
+  chain: string;
+  chainAddresses: Record<string, string>;
+  multiProvider: MultiProvider;
+  contractVerifier?: ContractVerifier;
+  warpConfig: HypTokenRouterConfig;
+  ismFactoryDeployer: HyperlaneProxyFactoryDeployer;
+}): Promise<IsmConfig | undefined> {
+  const { interchainSecurityModule } = warpConfig;
+  if (
+    !interchainSecurityModule ||
+    typeof interchainSecurityModule === 'string'
+  ) {
+    rootLogger.info(
+      `Config Ism is ${
+        !interchainSecurityModule ? 'empty' : interchainSecurityModule
+      }, skipping deployment.`,
+    );
+    return interchainSecurityModule;
+  }
+
+  rootLogger.info(`Loading registry factory addresses for ${chain}...`);
+  rootLogger.info(
+    `Creating ${interchainSecurityModule.type} ISM for token on ${chain} chain...`,
+  );
+  rootLogger.info(
+    `Finished creating ${interchainSecurityModule.type} ISM for token on ${chain} chain.`,
+  );
+
+  const evmIsmModule = await EvmIsmModule.create({
+    chain,
+    mailbox: chainAddresses.mailbox,
+    multiProvider,
+    proxyFactoryFactories: extractIsmAndHookFactoryAddresses(chainAddresses),
+    config: interchainSecurityModule,
+    ccipContractCache,
+    contractVerifier,
+  });
+  const { deployedIsm } = evmIsmModule.serialize();
+  return deployedIsm;
+}
+
+async function createWarpHook({
+  ccipContractCache,
+  chain,
+  chainAddresses,
+  multiProvider,
+  contractVerifier,
+  warpConfig,
+}: {
+  ccipContractCache: CCIPContractCache;
+  chain: string;
+  chainAddresses: Record<string, string>;
+  multiProvider: MultiProvider;
+  contractVerifier?: ContractVerifier;
+  warpConfig: HypTokenRouterConfig;
+  ismFactoryDeployer: HyperlaneProxyFactoryDeployer;
+}): Promise<HookConfig | undefined> {
+  const { hook } = warpConfig;
+
+  if (!hook || typeof hook === 'string') {
+    rootLogger.info(
+      `Config Hook is ${!hook ? 'empty' : hook}, skipping deployment.`,
+    );
+    return hook;
+  }
+
+  rootLogger.info(`Loading registry factory addresses for ${chain}...`);
+  rootLogger.info(`Creating ${hook.type} Hook for token on ${chain} chain...`);
+
+  // If config.proxyadmin.address exists, then use that. otherwise deploy a new proxyAdmin
+  const proxyAdminAddress: Address =
+    warpConfig.proxyAdmin?.address ??
+    (await multiProvider.handleDeploy(chain, new ProxyAdmin__factory(), []))
+      .address;
+
+  const evmHookModule = await EvmHookModule.create({
+    chain,
+    multiProvider,
+    coreAddresses: {
+      mailbox: chainAddresses.mailbox,
+      proxyAdmin: proxyAdminAddress,
+    },
+    config: hook,
+    ccipContractCache,
+    contractVerifier,
+    proxyFactoryFactories: extractIsmAndHookFactoryAddresses(chainAddresses),
+  });
+  rootLogger.info(
+    `Finished creating ${hook.type} Hook for token on ${chain} chain.`,
+  );
+
+  const { deployedHook } = evmHookModule.serialize();
+  return deployedHook;
+}
